# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Play9ja Gaming Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /withdrawal:
    get:
      operationId: listWithdrawals
      tags: [withdrawal]
      summary: List user withdrawals
      responses:
        "200":
          description: Withdrawals list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Withdrawal"
    post:
      operationId: createWithdrawal
      tags: [withdrawal]
      summary: Submit a withdrawal request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalInput"
      responses:
        "201":
          description: Withdrawal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
        "403":
          description: Membership required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # MEMBERSHIP
...
  /admin/withdrawals/{id}:
    patch:
      operationId: updateWithdrawalStatus
      tags: [admin]
      summary: Approve or decline a withdrawal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"

  # ADMIN - ACTIVITY MESSAGES
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Withdrawal:
      type: object
      required: [id, userId, amount, status, bankName, accountNumber, accountName, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
        status:
          type: string
          enum: [pending, approved, declined]
        bankName:
          type: string
        accountNumber:
          type: string
        accountName:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WithdrawalInput:
      type: object
      required: [amount, bankName, accountNumber, accountName]
      properties:
        amount:
          type: number
          minimum: 500
        bankName:
          type: string
        accountNumber:
          type: string
        accountName:
          type: string
...
    StatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, declined]
        notes:
          type: ["string", "null"]
...